chore: reorganize frontend folder structure
