+++ apps/backend/src/core/graphql/schema.gql
@@ -3,33 +3,33 @@
 # ------------------------------------------------------
 
 type Category {
-	createdAt: DateTime!
-	id: ID!
-	imageUrl: String
-	name: String!
-	updatedAt: DateTime!
+  createdAt: DateTime!
+  id: ID!
+  imageUrl: String
+  name: String!
+  updatedAt: DateTime!
 }
 
 input CreateCategoryInput {
-	name: String!
+  name: String!
 }
 
 input CreateProductInput {
-	categoryId: String!
-	description: String
-	name: String!
-	price: Float!
-	slug: String!
+  categoryId: String!
+  description: String
+  name: String!
+  price: Float!
+  slug: String!
 }
 
 input CreateUserInput {
-	country: String!
-	email: String!
-	isSeller: Boolean! = false
-	name: String!
-	surname: String!
-	telegramUsername: String
-	walletAddress: String!
+  country: String!
+  email: String!
+  isSeller: Boolean! = false
+  name: String!
+  surname: String!
+  telegramUsername: String
+  walletAddress: String!
 }
 
 """
@@ -38,66 +38,66 @@ A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date
 scalar DateTime
 
 type Mutation {
-	createCategory(data: CreateCategoryInput!): Category!
-	createProduct(data: CreateProductInput!): ProductDTO!
-	createProductImage(createProductImage: ProductImageDTO!): ProductImage!
-	createUser(data: CreateUserInput!): User!
-	updateUser(data: UpdateUserInput!, walletAddress: String!): User!
+  createCategory(data: CreateCategoryInput!): Category!
+  createProduct(data: CreateProductInput!): ProductDTO!
+  createProductImage(createProductImage: ProductImageDTO!): ProductImage!
+  createUser(data: CreateUserInput!): User!
+  updateUser(data: UpdateUserInput!, walletAddress: String!): User!
 }
 
 type ProductDTO {
-	categoryId: String!
-	createdAt: DateTime!
-	description: String
-	id: ID!
-	name: String!
-	price: Float!
-	slug: String!
-	updatedAt: DateTime!
+  categoryId: String!
+  createdAt: DateTime!
+  description: String
+  id: ID!
+  name: String!
+  price: Float!
+  slug: String!
+  updatedAt: DateTime!
 }
 
 type ProductImage {
-	createdAt: DateTime!
-	id: ID!
-	imageUrl: String!
-	productId: String!
-	updatedAt: DateTime!
+  createdAt: DateTime!
+  id: ID!
+  imageUrl: String!
+  productId: String!
+  updatedAt: DateTime!
 }
 
 input ProductImageDTO {
-	imageUrl: String!
-	productId: String!
+  imageUrl: String!
+  productId: String!
 }
 
 type Query {
-	categories: [Category!]!
-	category(id: String!): Category
-	product(id: String!): ProductDTO
-	productImage(id: String!): ProductImage
-	productImages: [ProductImage!]!
-	products: [ProductDTO!]!
-	user(walletAddress: String!): User
-	users: [User!]!
+  categories: [Category!]!
+  category(id: String!): Category
+  product(id: String!): ProductDTO
+  productImage(id: String!): ProductImage
+  productImages: [ProductImage!]!
+  products: [ProductDTO!]!
+  user(walletAddress: String!): User
+  users: [User!]!
 }
 
 input UpdateUserInput {
-	country: String
-	email: String
-	isSeller: Boolean = false
-	name: String
-	surname: String
-	telegramUsername: String
-	walletAddress: String
+  country: String
+  email: String
+  isSeller: Boolean = false
+  name: String
+  surname: String
+  telegramUsername: String
+  walletAddress: String
 }
 
 type User {
-	country: String!
-	createdAt: DateTime!
-	email: String!
-	isSeller: Boolean!
-	name: String!
-	surname: String!
-	telegramUsername: String
-	updatedAt: DateTime!
-	walletAddress: String!
+  country: String!
+  createdAt: DateTime!
+  email: String!
+  isSeller: Boolean!
+  name: String!
+  surname: String!
+  telegramUsername: String
+  updatedAt: DateTime!
+  walletAddress: String!
 }
