+++ apps/frontend/app/(routes)/seller/sales/[id]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import SaleDetailPage from "@/components/seller/ui/pages/SaleDetail";
+import SaleDetailPage from "@/components/seller/sale-detail";
 
 export default function SalesDetailsPage() {
 	return <SaleDetailPage />;
