+++ apps/frontend/app/(routes)/shopping/page.tsx
@@ -1,4 +1,4 @@
-import ShoppingTable from "./shopping-table";
+import ShoppingTable from "@/components/shopping/shopping-table";
 
 export default function ShoppingPage() {
 	return <ShoppingTable />;
