+++ apps/frontend/components/seller/Chat.tsx
@@ -5,7 +5,7 @@ import { Input } from "@/components/ui/input";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { Send } from "lucide-react";
 import { useTranslations } from "next-intl";
-import { useChat } from "./hooks/chat.hook";
+import { useChat } from "../../hooks/chat.hook";
 
 export function ChatComponent() {
 	const { messages, newMessage, sendMessage, setNewMessage } = useChat();
