+++ apps/frontend/components/seller/analytics/category-chart.tsx
@@ -1,82 +1,90 @@
 "use client";
 
-import {
-	SalesByDay,
-	TimeRange,
-} from "@/components/seller/mock/sales-analytics.mock";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import {
+	CategorySales,
+	TimeRange,
+} from "@/lib/mocks/seller/sales-analytics.mock";
 import { useTranslations } from "next-intl";
 import {
-	Line,
-	LineChart,
+	Cell,
+	Pie,
+	PieChart,
 	TooltipProps as RechartsTooltipProps,
 	ResponsiveContainer,
 	Tooltip,
-	XAxis,
-	YAxis,
 } from "recharts";
 
-interface SalesChartProps {
-	data: SalesByDay[];
+interface CategoryChartProps {
+	data: CategorySales[];
 	timeRange: TimeRange;
 }
 
-export function SalesChart({ data, timeRange }: SalesChartProps) {
+interface LabelProps {
+	name: string;
+	value: number;
+}
+
+export function CategoryChart({ data, timeRange }: CategoryChartProps) {
 	const t = useTranslations();
 	const getTitle = () => {
 		return (
-			t(`Sales.analytics.salesChart.title.${timeRange}`) ||
-			t("Sales.analytics.salesChart.title.default")
+			t(`Sales.analytics.categoryChart.title.${timeRange}`) ||
+			t("Sales.analytics.categoryChart.title.default")
 		);
 	};
 
 	return (
-		<Card className="col-span-4">
+		<Card className="col-span-2 mt-4 md:mt-0">
 			<CardHeader className="mb-4">
 				<CardTitle className="text-2xl font-semibold">{getTitle()}</CardTitle>
 				<p className="text-sm text-muted-foreground pl-4">
-					{t("Sales.analytics.salesChart.subtitle")}
+					{t("Sales.analytics.categoryChart.subtitle")}
 				</p>
 			</CardHeader>
 			<CardContent className="h-[300px]">
 				<ResponsiveContainer width="100%" height="100%">
-					<LineChart data={data}>
-						<XAxis
-							dataKey="date"
-							tickFormatter={(value: string) =>
-								new Date(value).getDate().toString()
-							}
-							stroke="#888888"
-							fontSize={12}
-						/>
-						<YAxis
-							stroke="#888888"
-							fontSize={12}
-							tickFormatter={(value: number) => `$${value}`}
-						/>
+					<PieChart>
+						<Pie
+							data={data}
+							dataKey="value"
+							nameKey="name"
+							cx="50%"
+							cy="50%"
+							outerRadius={80}
+							label={({ name, value }: LabelProps) => `${name}: ${value}%`}
+						>
+							{data.map((entry) => (
+								<Cell key={entry.name} fill={entry.color} />
+							))}
+						</Pie>
 						<Tooltip
 							content={({
 								active,
 								payload,
 							}: RechartsTooltipProps<number, string>) => {
 								if (active && payload && payload.length) {
-									const data = payload[0].payload as SalesByDay;
+									const data = payload[0].payload as CategorySales;
 									return (
 										<div className="rounded-lg border bg-background p-2 shadow-sm">
 											<div className="grid grid-cols-2 gap-2">
 												<div className="flex flex-col">
 													<span className="text-[0.70rem] uppercase text-muted-foreground">
-														{t("Sales.analytics.salesChart.tooltip.date")}
+														{t(
+															"Sales.analytics.categoryChart.tooltip.category",
+														)}
 													</span>
 													<span className="font-bold text-muted-foreground">
-														{new Date(data.date).toLocaleDateString()}
+														{data.name}
 													</span>
 												</div>
 												<div className="flex flex-col">
 													<span className="text-[0.70rem] uppercase text-muted-foreground">
-														{t("Sales.analytics.salesChart.tooltip.sales")}
+														{t(
+															"Sales.analytics.categoryChart.tooltip.percentage",
+														)}
 													</span>
-													<span className="font-bold">${data.sales}</span>
+													<span className="font-bold">{data.value}%</span>
 												</div>
 											</div>
 										</div>
@@ -85,19 +93,7 @@ export function SalesChart({ data, timeRange }: SalesChartProps) {
 								return null;
 							}}
 						/>
-						<Line
-							type="monotone"
-							dataKey="sales"
-							strokeWidth={2}
-							activeDot={{
-								r: 4,
-								style: { fill: "hsl(var(--primary))" },
-							}}
-							style={{
-								stroke: "hsl(var(--primary))",
-							}}
-						/>
-					</LineChart>
+					</PieChart>
 				</ResponsiveContainer>
 			</CardContent>
 		</Card>
