+++ apps/frontend/components/shopping/chat-message.tsx
