+++ apps/frontend/lib/mocks/seller/chat.mock.ts
