+++ apps/frontend/lib/mocks/seller/sales-analytics.mock.ts
